pyproject.toml
[tool.poetry]
name = "lippy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "0.3.26"
google-api-python-client = "1.7.2"
google-auth = "1.8.0"
google-auth-httplib2 = "0.0.3"
google-auth-oauthlib = "0.4.1"
langchain = "0.0.227"
instructorembedding = "1.0.1"
transformers = "4.30.2"
sentence-transformers = "2.2.2"
unstructured = "0.8.0"
accelerate = "0.20.3"
einops = "0.6.1"
omegaconf = "^2.3.0"
openai = "^0.27.8"
xformers = "0.0.20"
llvmlite = "^0.40.1"
numba = "^0.57.1"
openai-whisper = "^20230314"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
